+++ src/netius/base/server.py
@@ -484,7 +484,9 @@ class DatagramServer(Server):
                 if not error_v in SSL_VALID_ERRORS
                 and hasattr(error, "reason")
                 and error.reason
-                else None
+                else (
+                    self._ssl_reason(error) if not error_v in SSL_VALID_ERRORS else None
+                )
             )
             if error_v in SSL_SILENT_ERRORS or error_m in SSL_SILENT_REASONS:
                 self.on_expected(error)
@@ -522,7 +524,9 @@ class DatagramServer(Server):
                 if not error_v in SSL_VALID_ERRORS
                 and hasattr(error, "reason")
                 and error.reason
-                else None
+                else (
+                    self._ssl_reason(error) if not error_v in SSL_VALID_ERRORS else None
+                )
             )
             if error_v in SSL_SILENT_ERRORS or error_m in SSL_SILENT_REASONS:
                 self.on_expected(error)
@@ -767,7 +771,9 @@ class StreamServer(Server):
                 if not error_v in SSL_VALID_ERRORS
                 and hasattr(error, "reason")
                 and error.reason
-                else None
+                else (
+                    self._ssl_reason(error) if not error_v in SSL_VALID_ERRORS else None
+                )
             )
             if error_v in SSL_SILENT_ERRORS or error_m in SSL_SILENT_REASONS:
                 self.on_expected_s(error)
@@ -851,7 +857,9 @@ class StreamServer(Server):
                 if not error_v in SSL_VALID_ERRORS
                 and hasattr(error, "reason")
                 and error.reason
-                else None
+                else (
+                    self._ssl_reason(error) if not error_v in SSL_VALID_ERRORS else None
+                )
             )
             if error_v in SSL_SILENT_ERRORS or error_m in SSL_SILENT_REASONS:
                 self.on_expected(error, connection)
@@ -896,7 +904,9 @@ class StreamServer(Server):
                 if not error_v in SSL_VALID_ERRORS
                 and hasattr(error, "reason")
                 and error.reason
-                else None
+                else (
+                    self._ssl_reason(error) if not error_v in SSL_VALID_ERRORS else None
+                )
             )
             if error_v in SSL_SILENT_ERRORS or error_m in SSL_SILENT_REASONS:
                 self.on_expected(error, connection)